[fix] update MainActivity/color/theme
