+++ app/src/main/java/com/codaid/trailogandroid/SettingActivity.kt
@@ -1,0 +1,69 @@
+package com.codaid.trailogandroid
+
+import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.GravityCompat
+import com.codaid.trailogandroid.common.Utils
+import com.codaid.trailogandroid.common.Utils.Companion.navList
+import com.codaid.trailogandroid.common.Utils.Companion.navTitles
+import com.codaid.trailogandroid.common.Utils.Companion.optionList
+import com.codaid.trailogandroid.databinding.ActivitySettingBinding
+import com.google.android.material.navigation.NavigationView
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
+import kotlin.properties.Delegates
+
+class SettingActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
+
+    private lateinit var binding: ActivitySettingBinding
+    private var mGenre by Delegates.notNull<Int>()
+    private val utils = Utils()
+    private val auth = Firebase.auth
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        mGenre = 4
+        binding = ActivitySettingBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        setSupportActionBar(binding.appBar.toolbar)
+        utils.createToolbar(
+            this,
+            supportActionBar,
+            binding.navView,
+            binding.drawerLayout,
+            binding.appBar.toolbar
+        )
+        binding.appBar.toolbar.title = navTitles[mGenre]
+        binding.navView.setNavigationItemSelectedListener(this)
+        binding.appBar.contentSetting.logout.setOnClickListener {
+            auth.signOut()
+        }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        mGenre = 4
+        utils.selectItem(binding.navView, mGenre)
+        binding.appBar.toolbar.title = navTitles[mGenre]
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        menuInflater.inflate(R.menu.menu_main, menu)
+        return true
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        super.onOptionsItemSelected(item)
+        utils.goAnotherActivity(binding.appBar.toolbar, optionList.indexOf(item.itemId), "option")
+        return true
+    }
+
+    override fun onNavigationItemSelected(item: MenuItem): Boolean {
+        utils.goAnotherActivity(binding.appBar.toolbar, navList.indexOf(item.itemId), "nav")
+        binding.drawerLayout.closeDrawer(GravityCompat.START)
+        return true
+    }
+
+}
