+++ app/src/main/java/com/codaid/trailogandroid/SignupActivity.kt
@@ -8,7 +8,6 @@ import android.view.inputmethod.InputMethodManager
 import androidx.appcompat.app.AppCompatActivity
 import com.codaid.trailogandroid.common.Utils
 import com.codaid.trailogandroid.databinding.ActivitySignupBinding
-import com.codaid.trailogandroid.main.dash_board.MainActivity
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseAuthException
