+++ app/src/main/java/com/codaid/trailogandroid/common/custom_model/Training.kt
@@ -3,7 +3,7 @@ package com.codaid.trailogandroid.common.custom_model
 import com.google.firebase.firestore.ServerTimestamp
 import java.util.*
 
-data class Training (
+data class Training(
     val contents: MutableMap<String, Map<String, String>>? = null,
     @ServerTimestamp
     val createdAt: Date? = null
