+++ app/src/main/java/com/codaid/trailogandroid/common/custom_model/Weight.kt
@@ -3,7 +3,7 @@ package com.codaid.trailogandroid.common.custom_model
 import com.google.firebase.firestore.ServerTimestamp
 import java.util.*
 
-data class Weight (
+data class Weight(
     val weight: Float? = null,
     @ServerTimestamp
     val createdAt: Date? = null
