--- app/src/main/java/com/codaid/trailogandroid/common/view_model/Result.kt
@@ -1,14 +1,0 @@
-package com.codaid.trailogandroid.common.view_model
-
-sealed class Result<out T : Any> {
-
-    data class Success<out T : Any>(val data: T) : Result<T>()
-    data class Error(val errorCode: String) : Result<Nothing>()
-
-    override fun toString(): String {
-        return when (this) {
-            is Success<*> -> "Success[data=$data]"
-            is Error -> "Error[exception=$errorCode]"
-        }
-    }
-}
