--- app/src/main/java/com/codaid/trailogandroid/common/view_model/SuccessError.kt
@@ -1,8 +1,0 @@
-package com.codaid.trailogandroid.common.view_model
-
-import com.google.firebase.auth.FirebaseUser
-
-data class SuccessError (
-    val success: FirebaseUser? = null,
-    val error: Int? = null
-)
