+++ app/src/main/java/com/codaid/trailogandroid/main/add_training/AddTrainingActivity.kt
@@ -15,18 +15,16 @@ import androidx.annotation.StringRes
 import androidx.appcompat.app.ActionBarDrawerToggle
 import androidx.core.content.ContextCompat
 import androidx.core.view.GravityCompat
-import androidx.core.view.children
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.codaid.trailogandroid.databinding.ActivityAddTrainingBinding
 import com.codaid.trailogandroid.main.add_workout.AddWorkoutActivity
-import com.codaid.trailogandroid.main.dash_board.MainActivity
-import com.codaid.trailogandroid.main.setting.SettingActivity
+import com.codaid.trailogandroid.MainActivity
+import com.codaid.trailogandroid.SettingActivity
 import com.codaid.trailogandroid.main.add_weight.AddWeightActivity
 import com.google.android.material.navigation.NavigationView
 import kotlin.properties.Delegates
 import com.codaid.trailogandroid.R
-import com.codaid.trailogandroid.main.add_weight.afterTextChanged
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
