+++ app/src/main/res/layouts/activity/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
-    tools:context=".main.dash_board.MainActivity">
+    tools:context=".MainActivity">
 
     <include
         layout="@layout/app_bar_main"
