+++ app/src/main/res/layouts/activity/layout/activity_setting.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:openDrawer="start"
-    tools:context=".main.setting.SettingActivity">
+    tools:context=".SettingActivity">
 
     <include
         layout="@layout/app_bar_setting"
