+++ app/src/main/res/layouts/content/layout/content_setting.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.setting.SettingActivity">
+    tools:context=".SettingActivity">
 
     <Button
         android:id="@+id/logout"
